Estructura de clases

+++ Taller1_Hambuguesas/src/uniandes/dpoo/modelo/Producto.java
@@ -1,7 +1,7 @@
 package uniandes.dpoo.modelo;
 
 interface Producto {
-	public int getPrecio();
+	public double getPrecio();
 	public String getNombre();
 	public String generarTextoFactura();
 }
